@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": ">=7.1.0"
+        "php": ">=8.0.0"
     },
     "autoload": {
         "psr-4": {
